Cleaned up FileFragment
FileFragment was renamed to FilesFragment. It and its presenter are now organized and some bugs were fixed. The refresh indicator now stops automatically after new files are loaded.

InternalFilesUtil is now being injected.
BaseActivity and BaseFragment provide reactive alternatives for startActivityForResult() and requestPermissions() as well as a convenience method restartService().